+++ include/utils.h
@@ -1,0 +1,6 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+void terminarPrograma(char mensaje[]);
+
+#endif
